Chore: address PR comments
