+++ src/api/Base.js
@@ -16,7 +16,10 @@ import {
     HTTP_POST,
     HTTP_PUT,
     HTTP_DELETE,
+    PLACEHOLDER_USER,
 } from '../constants';
+import set from 'lodash/set';
+import { USER_FIELDS } from '../util/fields';
 
 class Base {
     /**
@@ -392,6 +395,27 @@ class Base {
         } catch (error) {
             this.errorHandler(error);
         }
+    }
+
+    /**
+     * Fill user properties that are null in an object
+     *
+     * @param {Object} obj - some object
+     * @return {Object} new object with user placeholder
+     */
+    fillUserPlaceholder(obj: Object): Object {
+        const newObj = { ...obj };
+
+        USER_FIELDS.forEach((field: string) => {
+            if (
+                Object.prototype.hasOwnProperty.call(newObj, field) &&
+                newObj[field] === null
+            ) {
+                set(newObj, field, PLACEHOLDER_USER);
+            }
+        });
+
+        return newObj;
     }
 }
