+++ src/util/__tests__/fields-test.js
@@ -9,7 +9,6 @@ import {
     USER_FIELDS,
     findMissingProperties,
     fillMissingProperties,
-    fillUserPlaceholder,
 } from '../fields';
 import {
     FIELD_ID,
@@ -50,7 +49,6 @@ import {
     FIELD_ASSIGNED_TO,
     FIELD_RESOLUTION_STATE,
     FIELD_RESTORED_FROM,
-    PLACEHOLDER_USER,
 } from '../../constants';
 
 describe('util/fields', () => {
@@ -213,43 +211,6 @@ describe('util/fields', () => {
             expect(fillMissingProperties(obj, properties)).toEqual({
                 foo: { bar: 1, baz: { bum: null, bup: null } },
                 bar: { bum: null, bop: { bip: null } },
-            });
-        });
-    });
-
-    describe('fillUserPlaceholder()', () => {
-        test('should return same object if none of the user properties are present', () => {
-            const obj = { foo: 1 };
-            expect(fillUserPlaceholder(obj)).toEqual(obj);
-        });
-
-        test('should not modify existing created_by if the value is not null', () => {
-            const obj = { foo: 1, created_by: { bar: 2 } };
-            expect(fillUserPlaceholder(obj)).toEqual(obj);
-        });
-
-        test('should add placeholder user for created_by if value is null', () => {
-            const obj = { foo: 1, created_by: null };
-            expect(fillUserPlaceholder(obj)).toEqual({
-                foo: 1,
-                created_by: PLACEHOLDER_USER,
-            });
-        });
-
-        test('should add placeholder user for all user properties if values are null', () => {
-            const obj = {
-                foo: 1,
-                created_by: null,
-                owned_by: null,
-                assigned_to: null,
-                modified_by: null,
-            };
-            expect(fillUserPlaceholder(obj)).toEqual({
-                foo: 1,
-                created_by: PLACEHOLDER_USER,
-                owned_by: PLACEHOLDER_USER,
-                assigned_to: PLACEHOLDER_USER,
-                modified_by: PLACEHOLDER_USER,
             });
         });
     });
